+++ briefly/settings.py
@@ -15,6 +15,7 @@ from pathlib import Path
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
+TEMPLATE_DIR = os.path.join(BASE_DIR, '/templates')
 
 # Static files (CSS, JavaScript, Images)
 STATIC_URL = 'static/'
@@ -61,7 +62,7 @@ MIDDLEWARE = [
 ]
 
 ROOT_URLCONF = 'briefly.urls'
-TEMPLATE_DIR = os.path.join(BASE_DIR, 'templates')
+TEMPLATE_DIR = os.path.join(BASE_DIR, TEMPLATE_DIR)
 
 TEMPLATES = [
     {
