+++ briefly_app/templates/briefly_navbar.html
@@ -1,3 +1,3 @@
 <div class="navbar">
-    <a href=""><h1>Briefly.</h1></a>
+  <a href="/"><h1>Briefly.</h1></a>
 </div>
