+++ briefly_app/templates/template_category.html
@@ -1,74 +1,74 @@
-{% extends "./briefly_layout.html" %}
-{% load static %}
-{% block body %}
+{% extends "./briefly_layout.html" %} {% load static %} {% block body %}
 <div class="category">
-    <h3>Set Your Category</h3>
-    <div class="category--form">
-        <div class="category--form--input">
-            <input type="checkbox" id="business">
-            <label class="category-label" for="business">Business</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="politics">
-            <label class="category-label" for="politics">Politics</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="technology">
-            <label class="category-label" for="technology">Technology</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="education">
-            <label class="category-label" for="education">Education</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="sports">
-            <label class="category-label" for="sports">Sports</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="forestry">
-            <label class="category-label" for="forestry">Forestry</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="forestry">
-            <label class="category-label" for="forestry">Forestry</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="health">
-            <label class="category-label" for="health">Health</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="entertainment">
-            <label class="category-label" for="entertainment">Entertainment</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="science">
-            <label class="category-label" for="science">Science</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="travel">
-            <label class="category-label" for="travel">Travel</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="food">
-            <label class="category-label" for="food">Food</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="fashion">
-            <label class="category-label" for="fashion">Fashion</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="finance">
-            <label class="category-label" for="finance">Finance</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="lifestyle">
-            <label class="category-label" for="lifestyle">Lifestyle</label>
-        </div>
-        <div class="category--form--input">
-            <input type="checkbox" id="news">
-            <label class="category-label" for="news">News</label>
-        </div>
+  <h3>Set Your Category</h3>
+  <div class="category--form">
+    <div class="category--form--input">
+      <input type="checkbox" id="business" />
+      <label class="category-label" for="business">Business</label>
     </div>
-    <input class="button" type="button" value="save">    
+    <div class="category--form--input">
+      <input type="checkbox" id="politics" />
+      <label class="category-label" for="politics">Politics</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="technology" />
+      <label class="category-label" for="technology">Technology</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="education" />
+      <label class="category-label" for="education">Education</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="sports" />
+      <label class="category-label" for="sports">Sports</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="forestry" />
+      <label class="category-label" for="forestry">Forestry</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="forestry" />
+      <label class="category-label" for="forestry">Forestry</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="health" />
+      <label class="category-label" for="health">Health</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="entertainment" />
+      <label class="category-label" for="entertainment">Entertainment</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="science" />
+      <label class="category-label" for="science">Science</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="travel" />
+      <label class="category-label" for="travel">Travel</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="food" />
+      <label class="category-label" for="food">Food</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="fashion" />
+      <label class="category-label" for="fashion">Fashion</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="finance" />
+      <label class="category-label" for="finance">Finance</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="lifestyle" />
+      <label class="category-label" for="lifestyle">Lifestyle</label>
+    </div>
+    <div class="category--form--input">
+      <input type="checkbox" id="news" />
+      <label class="category-label" for="news">News</label>
+    </div>
+  </div>
+  <a href="/template_headlines">
+    <input class="button" type="button" value="save" />
+  </a>
 </div>
-{% endblock %} 
+{% endblock %}
