+++ briefly_app/templates/template_signup.html
@@ -1,0 +1,114 @@
+{% extends "./briefly_layout.html" %} {% load static %} {% block body %}
+<div class="login">
+  <h3>Sign up</h3>
+  <div class="login--form">
+    <div class="login--form--input">
+      <label for="name">First Name</label>
+      <input
+        type="text"
+        id="firstname"
+        name="firstname"
+        placeholder="Enter your first name"
+      />
+    </div>
+    <div class="login--form--input">
+      <label for="name">Last Name</label>
+      <input
+        type="text"
+        id="lastname"
+        name="lastname"
+        placeholder="Enter your last name"
+      />
+    </div>
+    <div class="login--form--input">
+      <label for="email">Email</label>
+      <input
+        type="text"
+        id="email"
+        name="email"
+        placeholder="Enter your email"
+      />
+    </div>
+    <div class="login--form--input">
+      <label for="country">Country</label>
+      <label class="selectbox">
+        <select>
+          <option value="Afghanistan">Afghanistan</option>
+          <option value="Albania">Albania</option>
+          <option value="Algeria">Algeria</option>
+          <option value="Andorra">Andorra</option>
+          <option value="Angola">Angola</option>
+          <option value="Antigua and Barbuda">Antigua and Barbuda</option>
+          <option value="Argentina">Argentina</option>
+          <option value="Armenia">Armenia</option>
+          <option value="Australia">Australia</option>
+          <option value="Austria">Austria</option>
+          <option value="Azerbaijan">Azerbaijan</option>
+          <option value="Bahamas">Bahamas</option>
+          <option value="Bahrain">Bahrain</option>
+          <option value="Bangladesh">Bangladesh</option>
+          <option value="Barbados">Barbados</option>
+          <option value="Belarus">Belarus</option>
+          <option value="Belgium">Belgium</option>
+          <option value="Belize">Belize</option>
+          <option value="Benin">Benin</option>
+          <option value="Bhutan">Bhutan</option>
+          <option value="Bolivia">Bolivia</option>
+          <option value="Bosnia and Herzegovina">Bosnia and Herzegovina</option>
+          <option value="Botswana">Botswana</option>
+          <option value="Brazil">Brazil</option>
+          <option value="Brunei">Brunei</option>
+          <option value="Bulgaria">Bulgaria</option>
+          <option value="Burkina Faso">Burkina Faso</option>
+          <option value="Burundi">Burundi</option>
+          <option value="Cabo Verde">Cabo Verde</option>
+          <option value="Cambodia">Cambodia</option>
+          <option value="Cameroon">Cameroon</option>
+          <option value="Canada">Canada</option>
+          <option value="Central African Republic">
+            Central African Republic
+          </option>
+          <option value="Chad">Chad</option>
+          <option value="Chile">Chile</option>
+          <option value="China">China</option>
+          <option value="Colombia">Colombia</option>
+          <option value="Comoros">Comoros</option>
+          <option value="Congo, Democratic Republic of the">
+            Congo, Democratic Republic of the
+          </option>
+          <option value="Congo, Republic of the">Congo, Republic of the</option>
+          <option value="Costa Rica">Costa Rica</option>
+          <option value="Croatia">Croatia</option>
+          <option value="Cuba">Cuba</option>
+          <option value="Cyprus">Cyprus</option>
+          <option value="Czech Republic">Czech Republic</option>
+          <option value="Denmark">Denmark</option>
+          <option value="Djibouti">Djibouti</option>
+          <option value="Dominica">Dominica</option>
+        </select>
+      </label>
+    </div>
+    <div class="login--form--input">
+      <label for="password">Password</label>
+      <input
+        type="text"
+        id="password"
+        name="password"
+        placeholder="Enter your password"
+      />
+    </div>
+    <div class="login--form--input">
+      <label for="password">Password Confirmation</label>
+      <input
+        type="text"
+        id="password"
+        name="password"
+        placeholder="Enter your password"
+      />
+    </div>
+  </div>
+  <a href="/template_login">
+    <input class="button" type="button" value="save" />
+  </a>
+</div>
+{% endblock %}
